pyproject.toml
[project]
name = "tag-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.123.5",
    "fastapi-pagination>=0.15.3",
    "pre-commit>=4.5.0",
    "pycountry>=24.6.1",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]

[tool.ruff]
line-length = 120
exclude = [
    ".git",
    ".venv",
    ".env",
    "migrations",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
]
ignore = [
    "B008",   # Do not perform function call `Depends` in argument default
    "E402",   # imports at the top of the file
    "E501",   # line too long (handled by formatter)
    "E731",   # Do not assign a `lambda` expression, use a `def` - rule ignored
    "RUF005", # (*list_display, 'is_deleted') instead of concatenation list_display + ('is_deleted',) - rule ignored
    "RUF010", # f'Connection error: {e}' -> f'Connection error: {e!s}'
    "SIM105", # Use `contextlib.suppress(TypeError)` instead of `try`-`except`-`pass` - rule ignored
    "SIM108", # Use `getattr` with a default value instead of catching `AttributeError` - rule ignored
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "single"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
